Fixed for views

+++ similar_movies/views.py
@@ -14,8 +14,8 @@ def home():
     """ This is a main view which display form to search a movie or tv shows"""
     if request.method == 'POST':
         title = request.form['title']
-        type = request.form['type']
-        return redirect(url_for('views.list_similar_show', title=title, type=type))
+        show_type = request.form['type']
+        return redirect(url_for('views.list_similar_show', title=title, type=show_type))
     else:
         return render_template('home.html',
                                user=current_user)
@@ -26,8 +26,8 @@ def list_similar_show():
     """ This view allows to display list of similar movies or tv shows """
     try:
         title = request.args.get("title")
-        type = request.args.get("type")
-        similar_shows = Similar(title, type)
+        show_type = request.args.get("type")
+        similar_shows = Similar(title, show_type)
         try:
             return_similar_shows = similar_shows.return_similar_shows()
         except IndexError:
@@ -82,6 +82,9 @@ def save_show():
 def delete_show(id):
     """ This function allows to remove movie or tv show from list for login user """
     show = SavedMovies.query.get(id)
+    if show.user_id != current_user.id:
+        flash("You are not allowed to delete this show", category='error')
+        return redirect(url_for('auth.profile'))
     db.session.delete(show)
     db.session.commit()
     flash("Show deleted from your profile", category='success')
@@ -106,6 +109,9 @@ def save_watched_show():
 def delete_watched_show(id):
     """ This function allows to remove movie or tv show from watched list for login user """
     show = WatchedMovies.query.get(id)
+    if show.user_id != current_user.id:
+        flash("You are not allowed to delete this show", category='error')
+        return redirect(url_for('auth.profile'))
     db.session.delete(show)
     db.session.commit()
     flash("Show deleted from your watch history", category='success')
